renoir: overlay: Refine display cutout region
By inspecting screen_round_corner_top.png from stock
product/overlay/DevicesOverlay.apk, we can observe that the cutout has
a radius of 32px and begins 34px below the top border and 67px to the
left edge. The rectangle approximation of the cutout should have the
same amount of padding on top of the cutout as below according to the
pixel device trees ("This configuartion effects status bar's height
actually. Set its height as double y position of
@dimen/config_mainBuiltInDisplayCutout to make cutout center
vertically" [1]). Additionally, add 8px padding at the right side of
the cutout to avoid content being too close to the cutout. Below is a
sketch of the cutout drawn within the rectangle approximation.
___________
|         |
|    _    |  ___34px
|  /   \  |
| |     | |
|  \   /  |
|    ‾    |  ‾‾‾98px
|         |
‾‾‾‾‾‾‾‾‾‾‾  ‾‾‾108px
  |     | |
  67px  | |
        | |
        131px
          |
          139px

[1]: https://android.googlesource.com/device/google/caimito/+/refs/heads/main/caiman/overlay/frameworks/base/core/res/res/values/config.xml#197

Change-Id: I117a347bbaddc42185b31127733fa7ee5078d147